requirements.txt
# Django and REST framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.10
dj-database-url==2.1.0

# Channels for WebSockets
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
redis==5.0.1
django-redis==5.4.0
hiredis==2.2.3

# Utilities
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
uvicorn==0.24.0
whitenoise==6.6.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1
selenium==4.16.0
webdriver-manager==4.0.1